+++ packages/suite/src/views/wallet/dashboard/index.tsx
@@ -1,20 +1,14 @@
 import React from 'react';
-import { connect } from 'react-redux';
-import BigNumber from 'bignumber.js';
-import { bindActionCreators } from 'redux';
-// import { FormattedMessage } from 'react-intl';
+import NetworkGroup from './components/NetworkGroup';
+import { FormattedMessage } from 'react-intl';
 import styled from 'styled-components';
-
-import { H4, CoinLogo, colors, variables, Loader, Button } from '@trezor/components';
+import { H4, Loader, Icon, colors } from '@trezor/components';
 import WalletLayout from '@wallet-components/WalletLayout';
-// import l10nCommonMessages from '@wallet-views/messages';
 import { sortByCoin } from '@wallet-utils/accountUtils';
-import * as routerActions from '@suite-actions/routerActions';
-import * as discoveryActions from '@wallet-actions/discoveryActions';
 import { NETWORKS } from '@wallet-config';
-import { AppState, Dispatch } from '@suite-types';
 import { Network, Account } from '@wallet-types';
-// import l10nMessages from './index.messages';
+import { Props } from './Container';
+import messages from './index.messages';
 
 const Content = styled.div`
     flex: 1;
@@ -25,101 +19,29 @@ const Content = styled.div`
 const CardsWrapper = styled.div`
     display: flex;
     flex-wrap: wrap;
-    justify-content: space-evenly;
+    justify-content: flex-start;
 `;
 
-const Card = styled.div`
+const AddMoreCoins = styled.div`
+    display: flex;
     border: 1px solid ${colors.GRAY_LIGHT};
-    flex: 1;
     border-radius: 4px;
+    width: 180px;
+    height: 110px;
+    color: ${colors.TEXT_SECONDARY};
+    justify-content: center;
+    align-items: center;
     padding: 10px;
-    min-width: 180px;
     margin: 4px;
-`;
-
-const StyledCoinLogo = styled(CoinLogo)`
-    opacity: 0.7;
-    transition: opacity 0.2s ease-in-out;
     cursor: pointer;
 
     &:hover {
-        opacity: 1;
+        border: 1px solid ${colors.DIVIDER};
     }
 `;
 
-const Header = styled.div`
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-`;
-
-const Details = styled.div`
-    font-size: ${variables.FONT_SIZE.SMALL};
-    color: ${colors.TEXT_SECONDARY};
-    padding-top: 4px;
-    padding-left: 38px;
-`;
-
-const DetailItem = styled.div`
-    display: flex;
-    flex-direction: row;
-`;
-
-const Name = styled.div`
-    padding-left: 12px;
-    font-size: ${variables.FONT_SIZE.BIG};
-    color: ${colors.TEXT_PRIMARY};
-`;
-
-const mapStateToProps = (state: AppState) => ({
-    accounts: state.wallet.accounts,
-    discovery: state.wallet.discovery,
-    settings: state.wallet.settings,
-});
-
-const mapDispatchToProps = (dispatch: Dispatch) => ({
-    getDiscoveryForDevice: () => dispatch(discoveryActions.getDiscoveryForDevice()),
-    goto: bindActionCreators(routerActions.goto, dispatch),
-});
-
-export type Props = ReturnType<typeof mapStateToProps> & ReturnType<typeof mapDispatchToProps>;
-
-interface NetworkGroup {
-    accounts: Account[];
-    network: Network;
-}
-
-const NetworkGroup = ({ accounts, network }: NetworkGroup) => {
-    const totalBalance = accounts.reduce((prev, a) => {
-        const bn = new BigNumber(prev).plus(a.formattedBalance);
-        return bn.toString();
-    }, '0');
-    const totalTransactions = accounts.reduce((prev, a) => {
-        const bn = new BigNumber(prev).plus(a.history.total);
-        return bn.toString();
-    }, '0');
-    return (
-        <Card>
-            <Header>
-                <StyledCoinLogo symbol={network.symbol} size={26} />
-                <Name>{network.name}</Name>
-            </Header>
-            <Details>
-                <DetailItem>Accounts used: {accounts.length}</DetailItem>
-                <DetailItem>
-                    Total balance: {totalBalance} {network.symbol.toUpperCase()}
-                </DetailItem>
-                {network.networkType !== 'ripple' && (
-                    <DetailItem>Total transactions: {totalTransactions}</DetailItem>
-                )}
-            </Details>
-        </Card>
-    );
-};
-
-const StyledButton = styled(Button)`
-    margin: 16px 4px 0px 4px;
-    border: 1px solid ${colors.GRAY_LIGHT};
+const IconWrapper = styled.div`
+    padding-right: 5px;
 `;
 
 const Dashboard = (props: Props) => {
@@ -155,23 +77,22 @@ const Dashboard = (props: Props) => {
                             <NetworkGroup key={symbol} network={network} accounts={group[symbol]} />
                         );
                     })}
+                    {!isLoading && (
+                        <AddMoreCoins
+                            onClick={() => {
+                                props.goto('wallet-settings');
+                            }}
+                        >
+                            <IconWrapper>
+                                <Icon icon="PLUS" size={10} color={colors.TEXT_SECONDARY} />
+                            </IconWrapper>
+                            <FormattedMessage {...messages.TR_ADD_MORE_COINS} />
+                        </AddMoreCoins>
+                    )}
                 </CardsWrapper>
-                {!isLoading && (
-                    <StyledButton
-                        onClick={() => props.goto('wallet-settings')}
-                        icon="PLUS"
-                        align="left"
-                        variant="white"
-                    >
-                        Add more coins
-                    </StyledButton>
-                )}
             </Content>
         </WalletLayout>
     );
 };
 
-export default connect(
-    mapStateToProps,
-    mapDispatchToProps,
-)(Dashboard);
+export default Dashboard;
