+++ packages/suite/src/views/wallet/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
 
-import Dashboard from '@wallet-views/dashboard';
+import Dashboard from '@wallet-views/dashboard/Container';
 
 const Wallet = () => {
     return <Dashboard />;
